apm-be-main/apm-be-main/requirements.txt
# https://www.djangoproject.com/
django
Pillow

# https://www.django-rest-framework.org/
djangorestframework
markdown
django-filter

# https://django-rest-framework-simplejwt.readthedocs.io/en/latest/getting_started.html
djangorestframework-simplejwt
